MESSAGE_POLL_TIMEOUT should be enough for extra conf and StartServer() to complete
This is a UT fixup for a change in _AwaitServerMessages.

Note that https://github.com/ycm-core/ycmd/pull/1434#issuecomment-625642055 is not exactly correct together with #1434.
Quote:
> (...) Note that the java completer has a ton of setup to do before it gets to the super().StartServer().

It's true that it has some additional tasks (wiping out workspace) before it calls *super()._StartServerNoLock()*. But
PR 1434 doesn't address it - wiping workspace is called from JavaCompleter.StartServer() at which point _server_started
is already True, so the following part from that PR in _AwaitServerMessages():

-return self._initialize_event.is_set()
+return not self._server_started or self._initialize_event.is_set()

did not make it return True. In other words, #1434 seems to not address the JavaCompleter issue, but only the issue with
potentially long extra conf Settings(). However according to
https://github.com/ycm-core/ycmd/issues/1433#issuecomment-625404051 this should rather be solved by increasing
MESSAGE_POLL_TIMEOUT. Anyway, one of the first things we do from OnFileReadyToParse() when server is not healthy is
setting _server_started to True.